get pdf report by dates

+++ project-report/src/main/java/com/tsystems/javaschool/uberbahnreports/ReportControllerImpl.java
@@ -9,7 +9,9 @@ import javax.faces.bean.SessionScoped;
 
 import java.io.File;
 import java.io.IOException;
+import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.time.ZoneId;
 import java.util.*;
 
 @ManagedBean
@@ -24,6 +26,7 @@ public class ReportControllerImpl {
     private String password;
     private Date since;
     private Date until;
+    private String errMsg;
 
     public String getLogin() {
         return login;
@@ -57,19 +60,36 @@ public class ReportControllerImpl {
         this.until = until;
     }
 
+    public String getErrMsg() {
+        return errMsg;
+    }
 
-    public void downloadPDF() throws IOException {
-        Collection<TicketInfo> ticketInfos = reportService.getPurchasedTickets(getLogin(), getPassword(), LocalDateTime.now(), LocalDateTime.now());
-        File pdfFile = reportService.createPdf(ticketInfos, LocalDateTime.now().minusDays(60), LocalDateTime.now());
-        reportService.uploadPDFToClient(pdfFile);
-        try {
-            pdfFile.delete();
-        } catch (Exception ex) {
-            ex.printStackTrace();
+    public void setErrMsg(String errMsg) {
+        this.errMsg = errMsg;
+    }
+
+    public LocalDate convertDateToLocalDate (Date date){
+        return date.toInstant().atZone(ZoneId.systemDefault()).toLocalDate();
+    }
+
+    public void downloadPDF() {
+        LocalDate dateSince = convertDateToLocalDate(getSince());
+        LocalDate dateUntil = convertDateToLocalDate(getUntil());
+        Collection<Tickets> ticketInfos = reportService.getPurchasedTickets(getLogin(), getPassword(), dateSince, dateUntil);
+        if (ticketInfos == null) {
+            setErrMsg("Invalid login or password");
+        } else {
+            setErrMsg("");
+            File pdfFile = reportService.createPdf(ticketInfos, dateSince, dateUntil);
+            try {
+                reportService.uploadPDFToClient(pdfFile);
+                pdfFile.delete();
+            } catch (Exception ex) {
+                ex.printStackTrace();
+            }
         }
 
     }
-
 
 
 }
